refactor(auth): rename auth controller file and update imports
Update controller filename from 'controller.ts' to 'auth.controller.ts' for better clarity and consistency. All related imports in test files and routes have been updated to reflect this change.
